added getter moved execution to a separate class
