+++ src/main/java/ru/gb/voevodin/Main.java
@@ -1,7 +1,9 @@
 package ru.gb.voevodin;
 
-public class Main {
-    public static void main(String[] args) {
 
-    }
+
+import java.util.Random;
+
+public class  Main  {
+
 }
